Move api tests into the packages that they're testing
